+++ backend/controllers/formController.js
@@ -66,9 +66,18 @@ exports.updateForm = catchAsyncErrors(async (req, res, next) => {
     } else {
         newFormDescription = req.body.formDescription
     }
+    const filesAttached = form.formFiles
+    const fileLength= filesAttached.length
+    let arrayIds = []
+    for (let i = 0; i < fileLength; i++) {
+        arrayIds.push(filesAttached[i].filename) 
+      }
     if (req.files == null || req.files == ''){
         newFormFiles = form.formFiles
     }else{
+        if(arrayIds.length != 0){
+            cloudinary.api.delete_resources(arrayIds,{ resource_type: 'raw' })
+           }
         newFormFiles = req.files
     }
     let newFormsData ={
@@ -97,14 +106,16 @@ exports.deleteForm = catchAsyncErrors(async (req, res, next) => {
     if (!form) { return next(new ErrorHandler('Form Id does not exist')) }
     
     const filesAttached = form.formFiles
-    fileLength= filesAttached.length
+    const fileLength= filesAttached.length
     let arrayIds = []
     for (let i = 0; i < fileLength; i++) {
         arrayIds.push(filesAttached[i].filename) 
       }
-
-      cloudinary.api.delete_resources(arrayIds, 
-        { resource_type: 'raw' })
+      if(arrayIds.length != 0){
+        cloudinary.api.delete_resources(arrayIds, 
+            { resource_type: 'raw' })
+      }
+      
 
     await form.remove()
     res.status(200).json({
